Final v1.0

+++ app/Http/Controllers/CostController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Cost;
+use App\Models\CostDetail;
 use Illuminate\Http\Request;
 
 class CostController extends Controller
@@ -89,7 +90,10 @@ class CostController extends Controller
         if(!$cost) {
             return redirect()->route('costs.index')->with('error', 'Costo no encontrado.');
         }
-
+        $quotation = CostDetail::with('costs')->where('cost_id',$id)->first();
+        if($quotation){
+            return redirect()->route('costs.index')->with('error', 'Este costo esta asociado a una cotización.');
+        }
         $cost->delete();
 
         return redirect()->route('costs.index')->with('success', 'Costo actualizado correctamente.');
